Resolve relation filter value and refactor
