+++ packages/twenty-front/src/modules/context-store/hooks/useComputeContextStoreFilters.ts
@@ -1,0 +1,56 @@
+import { ContextStoreTargetedRecordsRule } from '@/context-store/states/contextStoreTargetedRecordsRuleComponentState';
+import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
+import { RecordGqlOperationFilter } from '@/object-record/graphql/types/RecordGqlOperationFilter';
+import { Filter } from '@/object-record/object-filter-dropdown/types/Filter';
+import { useComputeViewRecordGqlOperationFilter } from '@/object-record/record-filter/hooks/useComputeViewRecordGqlOperationFilter';
+import { makeAndFilterVariables } from '@/object-record/utils/makeAndFilterVariables';
+
+export const useComputeContextStoreFilters = () => {
+  const { computeViewRecordGqlOperationFilter } =
+    useComputeViewRecordGqlOperationFilter();
+
+  const computeContextStoreFilters = (
+    contextStoreTargetedRecordsRule: ContextStoreTargetedRecordsRule,
+    contextStoreFilters: Filter[],
+    objectMetadataItem: ObjectMetadataItem,
+  ) => {
+    let queryFilter: RecordGqlOperationFilter | undefined;
+
+    if (contextStoreTargetedRecordsRule.mode === 'exclusion') {
+      queryFilter = makeAndFilterVariables([
+        computeViewRecordGqlOperationFilter(
+          contextStoreFilters,
+          objectMetadataItem?.fields ?? [],
+          [],
+        ),
+        contextStoreTargetedRecordsRule.excludedRecordIds.length > 0
+          ? {
+              not: {
+                id: {
+                  in: contextStoreTargetedRecordsRule.excludedRecordIds,
+                },
+              },
+            }
+          : undefined,
+      ]);
+    }
+    if (contextStoreTargetedRecordsRule.mode === 'selection') {
+      queryFilter =
+        contextStoreTargetedRecordsRule.selectedRecordIds.length > 0
+          ? {
+              id: {
+                in: contextStoreTargetedRecordsRule.selectedRecordIds,
+              },
+            }
+          : computeViewRecordGqlOperationFilter(
+              contextStoreFilters,
+              objectMetadataItem?.fields ?? [],
+              [],
+            );
+    }
+
+    return queryFilter;
+  };
+
+  return { computeContextStoreFilters };
+};
