+++ packages/twenty-front/src/modules/object-record/select/hooks/useRecordsForSelect.ts
@@ -10,18 +10,19 @@ import { getObjectFilterFields } from '@/object-record/select/utils/getObjectFil
 import { makeAndFilterVariables } from '@/object-record/utils/makeAndFilterVariables';
 import { makeOrFilterVariables } from '@/object-record/utils/makeOrFilterVariables';
 import { OrderBy } from '@/types/OrderBy';
+import { useMapRelationViewFilterValueSpecialIdsToRecordIds } from '@/views/view-filter-value/hooks/useResolveRelationViewFilterValue';
 
 export const useRecordsForSelect = ({
   searchFilterText,
   sortOrder = 'AscNullsLast',
-  selectedIds,
+  selectedRecordIdsAndSpecialIds,
   limit,
   excludedRecordIds = [],
   objectNameSingular,
 }: {
   searchFilterText: string;
   sortOrder?: OrderBy;
-  selectedIds: string[];
+  selectedRecordIdsAndSpecialIds: string[];
   limit?: number;
   excludedRecordIds?: string[];
   objectNameSingular: string;
@@ -41,15 +42,22 @@ export const useRecordsForSelect = ({
     objectNameSingular,
   });
 
+  const { mapRelationViewFilterValueSpecialIdsToRecordIds } =
+    useMapRelationViewFilterValueSpecialIdsToRecordIds();
+
+  const selectedRecordIds = mapRelationViewFilterValueSpecialIdsToRecordIds(
+    selectedRecordIdsAndSpecialIds,
+  );
+
   const orderByField = getObjectOrderByField(sortOrder);
-  const selectedIdsFilter = { id: { in: selectedIds } };
+  const selectedIdsFilter = { id: { in: selectedRecordIds } };
 
   const { loading: selectedRecordsLoading, records: selectedRecordsData } =
     useFindManyRecords({
       filter: selectedIdsFilter,
       orderBy: orderByField,
       objectNameSingular,
-      skip: !selectedIds.length,
+      skip: !selectedRecordIds.length,
     });
 
   const searchFilters = filters.map(({ fieldNames, filter }) => {
@@ -88,10 +96,10 @@ export const useRecordsForSelect = ({
     filter: makeAndFilterVariables([...searchFilters, selectedIdsFilter]),
     orderBy: orderByField,
     objectNameSingular,
-    skip: !selectedIds.length,
+    skip: !selectedRecordIds.length,
   });
 
-  const notFilterIds = [...selectedIds, ...excludedRecordIds];
+  const notFilterIds = [...selectedRecordIds, ...excludedRecordIds];
   const notFilter = notFilterIds.length
     ? { not: { id: { in: notFilterIds } } }
     : undefined;
@@ -107,7 +115,7 @@ export const useRecordsForSelect = ({
     objectNameSingular === 'workspaceMember'
       ? [
           {
-            id: 'me',
+            id: 'CURRENT_WORKSPACE_MEMBER',
             name: 'Me',
             isSelected: false,
           },
