+++ packages/twenty-front/src/modules/views/hooks/useQueryVariablesFromActiveFieldsOfViewOrDefaultView.ts
@@ -1,7 +1,7 @@
 import { useActiveFieldMetadataItems } from '@/object-metadata/hooks/useActiveFieldMetadataItems';
 import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
+import { useGetQueryVariablesFromView } from '@/views/hooks/useGetQueryVariablesFromView';
 import { useViewOrDefaultViewFromPrefetchedViews } from '@/views/hooks/useViewOrDefaultViewFromPrefetchedViews';
-import { getQueryVariablesFromView } from '@/views/utils/getQueryVariablesFromView';
 import { useIsFeatureEnabled } from '@/workspace/hooks/useIsFeatureEnabled';
 
 export const useQueryVariablesFromActiveFieldsOfViewOrDefaultView = ({
@@ -11,6 +11,8 @@ export const useQueryVariablesFromActiveFieldsOfViewOrDefaultView = ({
   objectMetadataItem: ObjectMetadataItem;
   viewId: string | null | undefined;
 }) => {
+  const { getQueryVariablesFromView } = useGetQueryVariablesFromView();
+
   const { view } = useViewOrDefaultViewFromPrefetchedViews({
     objectMetadataItemId: objectMetadataItem.id,
     viewId,
