+++ packages/twenty-front/src/modules/views/view-filter-value/utils/resolveBooleanViewFilterValue.ts
@@ -1,7 +1,2 @@
-import { ViewFilter } from '@/views/types/ViewFilter';
-
-export const resolveBooleanViewFilterValue = (
-  viewFilter: Pick<ViewFilter, 'value'>,
-) => {
-  return viewFilter.value === 'true';
-};
+export const resolveBooleanViewFilterValue = (viewFilterValue: string) =>
+  viewFilterValue === 'true';
