--- packages/twenty-front/src/modules/views/view-filter-value/utils/resolveRelationViewFilterValue.ts
@@ -1,8 +1,0 @@
-import { ViewFilter } from '@/views/types/ViewFilter';
-
-export const resolveRelationViewFilterValue = (
-  viewFilter: Pick<ViewFilter, 'value'>,
-) => {
-  // TODO: convert 'CURRENT_USER' to the current user id
-  return viewFilter.value;
-};
